# Matmul performance improvements