- part of #23
- ~should also get started on #19 (still todo)~ future PR
- ~not sure when i'll get round to finishing this... hopefully in the next month or so~