@nickrobinson251 2019-10-18
I do not know when I will get to #19 (adding benchmarks should be easier then it is)

But here's a quick and dirty benchmark showing this PR is an improvement of some test cases:
```
for nblocks in (2, 20, 200)
           println(nblocks)
           B = BlockDiagonal([rand(10, 10) for _ in 1:nblocks]);
           A = rand(size(B)...);
           M = Matrix(B);

           println("BlockDiag * Matrix")
           @btime $B * $A;
           @btime $M * $A;

           println("Matrix * BlockDiag")
           @btime $A * $B;
           @btime $A * $M;
end
```
running this and combining the results:
```
2
BlockDiag * Matrix
  old    3.328 μs (24 allocations: 10.91 KiB)
  new    2.037 μs (11 allocations: 7.34 KiB)
  base   1.175 μs (1 allocation: 3.25 KiB)

Matrix * BlockDiag
  old   3.549 μs (32 allocations: 11.06 KiB)
  new   1.705 μs (11 allocations: 7.34 KiB)
  base  1.060 μs (1 allocation: 3.25 KiB)

20
BlockDiag * Matrix
  old    197.224 μs (187 allocations: 946.69 KiB)
  new    132.566 μs (48 allocations: 629.97 KiB)
  base   114.097 μs (2 allocations: 312.58 KiB)

Matrix * BlockDiag
  old   156.342 μs (267 allocations: 948.25 KiB)
  new   95.509 μs (48 allocations: 629.97 KiB)
  base  111.817 μs (2 allocations: 312.58 KiB)

200
BlockDiag * Matrix
  old   78.422 ms (2504 allocations: 91.63 MiB)
  new   39.232 ms (608 allocations: 61.07 MiB)
  base  72.363 ms (2 allocations: 30.52 MiB)

Matrix * BlockDiag
  old   56.001 ms (3304 allocations: 91.64 MiB)
  new   32.039 ms (608 allocations: 61.07 MiB)
  base  71.763 ms (2 allocations: 30.52 MiB)
```
where `old` is current master, `new` is this PR and `base` is using plain-old `Matrix` instead of `BlockDiagonal`. Basically we're still not great on not allocating, but we've made some progress on speed 

```
julia> versioninfo()
Julia Version 1.3.0-rc4.1
Commit 8c4656b97a (2019-10-15 14:08 UTC)
Platform Info:
  OS: macOS (x86_64-apple-darwin18.6.0)
  CPU: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-6.0.1 (ORCJIT, skylake)
```